requirements.txt
django==6.0
djangorestframework==3.16
requests==2.32
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.2.2
django-filter==24.3